+++ index.js
@@ -54,6 +54,9 @@ function createError () {
         break
       case 'number':
         status = arg
+        if (i !== 0) {
+          deprecate('non-fist-argument status code; replace with createError(' + arg + ', ...)')
+        }
         break
       case 'object':
         props = arg